pyproject.toml
[project]
name = "tanda_assistant"
version = "24.4.23"
description = "Lorem"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "panel",
  "stravalib",
  "altair[all]",
  "stravalib",
  "redis[hiredis]",
  "starlette-session"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "watchfiles",
  "ipykernel",
  "jupyter_black",
  "torch",
  "transformers",
  "outlines",
  "datasets",
]

[tool.ruff]
exclude = [".env", ".venv", "venv", "notebooks"]
line-length = 79

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "I", "W"]
fixable = ["I"]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
relative_files = true

[tool.coverage.report]
show_missing = true
fail_under = 80
